requirements.txt
fastapi
uvicorn
fastapi-limiter
redis
python-dotenv
pinecone
groq
PyPDF2